Share discovered item key parsing
